+++ starknet_py/conftest.py
@@ -6,4 +6,5 @@ pytest_plugins = [
     "starknet_py.tests.e2e.fixtures.contracts",
     "starknet_py.tests.e2e.client.fixtures.prepare_network",
     "starknet_py.tests.e2e.fixtures.utils",
+    "starknet_py.tests.e2e.fixtures.constants",
 ]
