+++ starknet_py/tests/e2e/declare/declare_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from starknet_py.tests.e2e.conftest import MAX_FEE
+from starknet_py.tests.e2e.fixtures.constants import MAX_FEE
 
 
 @pytest.mark.asyncio
