+++ starknet_py/tests/e2e/fixtures/constants.py
@@ -1,5 +1,3 @@
-# pylint: disable=redefined-outer-name
-
 import os
 from pathlib import Path
