+++ starknet_py/utils/typed_data_test.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 
-from starknet_py.tests.e2e.conftest import typed_data_dir
+from starknet_py.tests.e2e.fixtures.constants import typed_data_dir
 from starknet_py.utils.typed_data import TypedData, get_hex
 
 TD = "typed_data_example.json"
